added : to output

+++ python_introduction/finance_calculator.py
@@ -8,6 +8,6 @@ income = int(input("Enter your monthly income: "))
 expenses = int(input("Enter your total monthly expenses: "))
 
 savings = income - expenses
-projected_savings = float(savings * 12 + (savings * 12 * 0.05))
+projected_savings = (savings * 12) + (savings * 12 * 0.05)
 print(f"Your monthly savings is ${savings}.")
 print(f"Projected savings after one year, with interest, is: ${projected_savings:.0f}.")
